+++ lib/src/config/config.dart
@@ -3,4 +3,11 @@
 
 class Config {
   static final baseUrl = "http://student.daystar.ac.ke";
+
+  /// The generateHeader function creates headers from
+  /// necessary for the working of the school portal based on
+  /// the session
+  static Map<String, String> generateHeader(String session) {
+    return {'Cookie': session};
+  }
 }
